+++ docs/.vuepress/config.ts
@@ -4,6 +4,7 @@ import vue from "@vitejs/plugin-vue";
 import theme from "./theme.js";
 import { pwaPlugin } from "vuepress-plugin-pwa2";
 import { getDirname, path } from "@vuepress/utils";
+import { seoPlugin } from "vuepress-plugin-seo2";
 
 import { registerComponentsPlugin } from "@vuepress/plugin-register-components";
 
@@ -16,7 +17,7 @@ export default defineUserConfig({
   dest: "dist/docs",
 
   base: "/docs/",
-  title: "Welcome to our Docs | zkSync Era",
+  title: "Welcome to our Docs | zkSync Era Docs",
 
   head: [
     ["link", { rel: "canonical", href: "https://era.zksync.io/docs/" }],
@@ -50,7 +51,6 @@ export default defineUserConfig({
         content: "zkSync Era is a user-centric zk rollup platform from Matter Labs. It is a scaling solution for Ethereum, already live on Ethereum mainnet.",
       },
     ],
-
     ["meta", { name: "twitter:card", content: "summary" }],
     ["meta", { name: "twitter:title", content: "Welcome to our Docs | zkSync Era" }],
     [
@@ -99,7 +99,6 @@ export default defineUserConfig({
         let contributorsDiv = document.querySelector('.contributors');
         contributorsDiv.innerHTML = '<span class="label">Contributors: </span>' + updatedList;
       
-        console.log('Top 5 contributors:', updatedList);
       });
 
       `,
@@ -109,6 +108,7 @@ export default defineUserConfig({
   theme,
 
   plugins: [
+    seoPlugin,
     pwaPlugin({
       update: "force",
     }),